+++ lib/glimmer-dsl-web/samples/regular/todo_mvc/views/new_todo_input.rb
@@ -3,7 +3,7 @@ require_relative 'todo_input'
 class NewTodoInput < TodoInput
   option :presenter
   
-  markup { # evaluated against instance as a smart default convention
+  markup { # evaluated against instance as a smart convention
     input(placeholder: "What needs to be done?", autofocus: "") {
       value <=> [presenter.new_todo, :task]
     
@@ -13,10 +13,10 @@ class NewTodoInput < TodoInput
     }
   }
   
-  style { # evaluated against class as a smart default convention (common to all instances)
+  style { # evaluated against class as a smart convention (common to all instances)
     todo_input_styles
     
-    rule(".#{component_element_class}") { # built-in component_class.component_element_class (e.g. NewTodoInput has CSS class as new-todo-input)
+    rule(".#{component_element_class}") { # NewTodoInput has component_element_class as 'new-todo-input'
       padding '16px 16px 16px 60px'
       height '65px'
       border 'none'
@@ -24,7 +24,7 @@ class NewTodoInput < TodoInput
       box_shadow 'inset 0 -2px 1px rgba(0,0,0,0.03)'
     }
     
-    rule(".#{component_element_class}::placeholder") { # built-in component_class.component_element_class (e.g. NewTodoInput has CSS class as new-todo-input)
+    rule(".#{component_element_class}::placeholder") {
       font_style 'italic'
       font_weight '400'
       color 'rgba(0, 0, 0, 0.4)'